MOT-2529: Host crashes and restarts during CTS Sensor Test
MOT-2116: OSP 1.0 doesn't report Step Detect and Count

1. IIO device for STEP counter sensor was created without IIO buffer.
Step counter data was pushed by IIO event.
But in Middleware it was expecting data for step counter via IIO buffer.
So during CTS Sensor test, when sensor batch enable and flush command
is sent to host for step counter sensor, it crashes.

2. This patch contains fix for OSP step counter data doesnt report
to Android.

+++ ospd/OSPDaemon.h
@@ -124,7 +124,7 @@ static inline void conv2OSP(ASensorType_t sensor,
 		d->booldata.TimeStamp = ts;
 		break;
 	case SENSOR_STEP_COUNTER:
-		d->stepcount.StepCount = val[0];
+		d->stepcount.StepCount = (uint32_t)val[0];
 		d->stepcount.TimeStamp = ts;
 		break;
 	default:
